feat(config-ui): use new connections page to replace old
