+++ config-ui/src/routes/connection/index.ts
@@ -16,4 +16,4 @@
  *
  */
 
-export * from './detail';
+export * from './connections';
